- implement streamables and async data retrievals - reorganization of major files
