+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "oshconnect"
-version = "0.3.0a3.post3"
+version = "0.3.0a4"
 description = "Library for interfacing with OSH, helping guide visualization efforts, and providing a place to store configurations."
 readme = "README.md"
 authors = [
@@ -12,7 +12,8 @@ dependencies = [
     "pydantic>=2.7.4,<3.0.0",
     "shapely>=2.0.4,<3.0.0",
     "websockets>=12.0,<16.0",
-    "requests"
+    "requests",
+    "aiohttp>=3.12.15",
 ]
 
 [dependency-groups]
@@ -24,5 +25,4 @@ dev = [
 ]
 
 [tool.setuptools]
-#packages = ["oshconnect", "oshconnect.csapi4py", "oshconnect.datamodels", "oshconnect.csapi4py.comm"]
 packages = {find = { where = ["src/"]}}
