+++ src/oshconnect/__init__.py
@@ -6,4 +6,4 @@
 #   ==============================================================================
 
 from .oshconnectapi import OSHConnect
-from .osh_connect_datamodels import System, Node, Datastream, Observation, ControlChannel
+from .streamableresource import System, Node, Datastream, Observation, ControlChannel
