+++ src/oshconnect/csapi4py/con_sys_api.py
@@ -2,8 +2,8 @@ from typing import Union
 
 from pydantic import BaseModel, HttpUrl, Field
 
-from oshconnect.csapi4py.endpoints import Endpoint
-from oshconnect.csapi4py.request_wrappers import post_request, put_request, get_request, delete_request
+from .endpoints import Endpoint
+from .request_wrappers import post_request, put_request, get_request, delete_request
 
 
 class ConnectedSystemAPIRequest(BaseModel):
