--- src/oshconnect/csapi4py/part_1/__init__.py
@@ -1,7 +1,0 @@
-# import capabilities
-# import collections_ep
-# import deployments
-# import procedures
-# import properties
-# import sampling_features
-# import systems
