--- src/oshconnect/csapi4py/part_2/__init__.py
@@ -1,6 +1,0 @@
-# import commands
-# import control_channels
-# import datastreams
-# import observations
-# import system_events
-# import system_history
