--- src/oshconnect/datamodels/__init__.py
@@ -1,10 +1,0 @@
-# import api_utils
-# import commands
-# import control_streams
-# import datastreams
-# import encoding
-# import geometry
-# import network_properties
-# import observations
-# import swe_components
-# import system_events_and_history
