+++ src/oshconnect/geometry.py
@@ -1,6 +1,13 @@
+#  =============================================================================
+#  Copyright (c) 2025 Botts Innovative Research Inc.
+#  Date: 2025/10/1
+#  Author: Ian Patterson
+#  Contact Email: ian@botts-inc.com
+#  =============================================================================
+
 from pydantic import BaseModel, Field
 
-from oshconnect.csapi4py.constants import GeometryTypes
+from .csapi4py.constants import GeometryTypes
 
 
 # TODO: Add specific validations for each type
