+++ src/oshconnect/network_properties.py
